Added the MathJax to the site.
From a request by oscarvarto on the IRC channel, we have now included
MathJax as a library on the documentation site so that decent math
formulae etc can be displayed.

I had no idea about this library and am VERY HAPPY to be able to include
it. This solves some really complicated LaTeX integration that I have been
working on for a while now.

Signed-off-by: Gary Pamparà <gpampara@gmail.com>